+++ test/Header.test.js
@@ -140,20 +140,23 @@ describe('Header', () => {
         modelPatronName.restore();
       });
 
-      it('should call the function to check if the cookie "nyplIdentityPatron" exists', () => {
+      // Temporary (?) skip because cookie login is not supported in Vega.
+      it.skip('should call the function to check if the cookie "nyplIdentityPatron" exists', () => {
         expect(setLoginCookie.calledOnce).to.equal(true);
         expect(hasCookie.calledOnce).to.equal(true);
         expect(hasCookie.calledWith('nyplIdentityPatron')).to.equal(true);
       });
 
-      it('should call the function to get the value of "nyplIdentityPatron" cookie, ' +
+      // Temporary (?) skip because cookie login is not supported in Vega.
+      it.skip('should call the function to get the value of "nyplIdentityPatron" cookie, ' +
         'if the cookie exists', () => {
         expect(getNyplIdentityPatronCookie.calledOnce).to.equal(true);
         expect(getNyplIdentityPatronCookie.alwaysCalledWithExactly('nyplIdentityPatron'))
           .to.equal(true);
       });
 
-      it('should call the API endpoint to get logged in patron\'s data with the cookie we got',
+      // Temporary (?) skip because cookie login is not supported in Vega.
+      it.skip('should call the API endpoint to get logged in patron\'s data with the cookie we got',
         () => {
           expect(fetchPatronData.calledOnce).to.equal(true);
           expect(getPatronData.calledOnce).to.equal(true);
@@ -233,7 +236,7 @@ describe('Header', () => {
 
           getCookie = sinon.stub(utils, 'getCookie')
             .withArgs('VALID_DOMAIN_LAST_VISITED')
-            // Set a mock last visit time for Encore timer
+            // Set a mock last visit time for Encore Catalog timer
             // So cookie "nyplIdentityPatron" will be kept until we test it's access token
             .returns(Date.now());